+++ source/ESPEED32/telemetry_logging.h
@@ -48,7 +48,7 @@ typedef struct {
   uint16_t sensi_halfPct;
   uint8_t trigger_pct;
   uint8_t output_pct;
-  uint8_t brake_pct;
+  uint16_t brake_pct;      /* [0.1%] Applied brake value */
   uint8_t carIndex;
   uint8_t releaseMode;
   uint8_t flags;
@@ -110,7 +110,7 @@ void telemetryCaptureSample(uint8_t carIndex,
                             uint8_t outputPct,
                             uint16_t vin_mV,
                             uint16_t current_mA,
-                            uint8_t brakePct,
+                            uint16_t brakePct,
                             uint16_t sensi_halfPct,
                             uint8_t releaseMode,
                             uint8_t flags);
